cpu: x64: matmul: move brgemm reduce kernel out of jit_brgemm_post_ops
The reduce kernel is used only by brgemm matmul to implement the `reduce`
attribute.
